Refactor playlist script to make code cleaner
