+++ portfolio/src/main/webapp/index.html
@@ -8,7 +8,7 @@
     <link rel="apple-touch-icon" sizes="180x180" href="images/apple-touch-icon.png">
     <link rel="icon" type="image/png" sizes="32x32" href="images/favicon-32x32.png">
     <link rel="icon" type="image/png" sizes="16x16" href="images/favicon-16x16.png">
-    <script src="script.js"></script>
+    <script src="script.js" type="module"></script>
     <title>Ife Omidiran</title>
   </head>
   <body>
