+++ portfolio/src/main/webapp/playlists.html
@@ -8,7 +8,7 @@
     <link rel="apple-touch-icon" sizes="180x180" href="images/apple-touch-icon.png">
     <link rel="icon" type="image/png" sizes="32x32" href="images/favicon-32x32.png">
     <link rel="icon" type="image/png" sizes="16x16" href="images/favicon-16x16.png">
-    <script src="script.js"></script>
+    <script src="script.js" type="module"></script>
     <title>Playlists | Ife Omidiran </title>
   </head>
   <body>
@@ -70,7 +70,7 @@
             Please provide the playlist length.
           </div>
         </div>
-        <button onclick="validatePlaylistForm();" type="button" class="btn btn-dark mb-3"
+        <button type="button" class="btn btn-dark mb-3"
             id="playlistSubmit">
           Get your playlist!
         </button>
